Move the data parsing into helpers.js
